+++ 06/main.c
@@ -15,10 +15,10 @@ int main(int argc, char **argv){
 	char* input_file = argv[1];
 
 	int* data = read_row_of_ints(input_file, ',', 500);
-    long long* countdowns = parse_countdown(data, 500);
-    printf("Number of fish after 80 days: %lld\n",
+    long* countdowns = parse_countdown(data, 500);
+    printf("Number of fish after 80 days: %ld\n",
             count_after_n_days(countdowns, 80));
-    printf("Number of fish after 256 days: %lld\n",
+    printf("Number of fish after 256 days: %ld\n",
             count_after_n_days(countdowns, 256-80));
 
     free(countdowns);
